move view emitter to separate file
